refactor: SnapPage.tsx 경로 수정 & importing 절대 경로 적용
