+++ snapsplit/src/features/trip/[tripId]/snap/_components/PhotoGrid.tsx
