+++ snapsplit/src/features/trip/[tripId]/snap/_components/TabSelector.tsx
