+++ snapsplit/src/features/trip/[tripId]/snap/_components/TopBar.tsx
