+++ snapsplit/src/features/trip/[tripId]/snap/_components/UploadButton.tsx
