+++ snapsplit/src/features/trip/[tripId]/snap/_components/tabView/BaseTabView.tsx
