+++ snapsplit/src/features/trip/[tripId]/snap/type.ts
